A bit more cleanup for shorter files.
